git-svn-id: https://svn.apache.org/repos/asf/lucene/dev/branches/lucene6238@1663307 13f79535-47bb-0310-9956-ffa450edef68

+++ lucene/core/src/test/org/apache/lucene/util/TestTimSorter.java
@@ -129,7 +129,7 @@ public class TestTimSorter extends BaseSortTestCase {
       Y += X + 1;
     }
 
-    if(runningTotal + X <= length) {
+    if((long) runningTotal + X <= length) {
       runningTotal += X;
       generateWrongElem(X, minRun, runs);
     }
